refactor: extract local helper function
Extract helper functions for getting auth-information from UPMConfig objects into the upm-config module, so they are close to other  UPMConfig related functions.
